Change three .flac files into .ogg in the adult edition
When testing the create-release script, I discovered three .flac files
in the villager/trade folder in the adult edition.  Not only were these
files not .ogg, but they were also stereo, and had no mention in the
sounds.json file.  All my attempts to review the git history to find out
how (and when) this happened failed, mostly because I'm still kind of a
git n00b.

In this commit, I processed these three files into mono .ogg and
provided records in the sounds.json file so that they will actually be
played in-game.

This fixes issue #9