Slightly reduce the volume of the villager "ambient" sounds
Through watching players use this pack and keeping note of what they say
when doing so, I was able to determine that the ambient sounds are a bit
too loud.  In fact, JazzLikeOnion says he needs to turn down the audio
on "friendly creatures" so as not to drive himself insane when near the
villagers. This is the opposite of what I was going for.

I knocked the volume down by .2 for every "ambient" sound in all three
editions.  In my testing, this is fairly subtle, but it does cut down
a bit on the "wall of noise" in an extensive villager trading hall.

This commit implements issue #27