Reduce the duration of an overly-long trade sound
The phrase that begins "I'll have two number nines..." was more than 5
seconds long, which allowed a single villager to talk over herself. The
next trade sound would play while the end of the previous was still
playing. This struck me as a bug.

I eliminated some of the silence before "... and a large gay seggs," and
used a filter to change the tempo of the phrase without changing the
pitch.  The phrase is now short enough to not be "talked-over" by the
next phrase.

This fixes issue #8