Add the ability to generate packs that allow sounds from lower packs
This commit allows a user to choose whether a pack will be a standard
pack or an "addon" pack. Standard packs suppress sounds for each event
in lower packs.  Addon packs allow all sounds from lower packs.

In my earlier testing, I was able to show that by putting the "loaded"
edition on the bottom of the list and both "adult" and "uwu" addon packs
above it, that I could get sounds from all three packs with no side
effects.

This commit is an initial proof-of-concept commit and has not been
tested in a Minecraft world. My testing consisted of looking at the list
of files and the sounds.json file to determine if everything looked ok.
I am checking it in now just to preserve my work before refactoring.

The refactoring I intend to do will remove the requirement to manually
choose the pack type, because now that I've added this, in order to
create the entire set of releases, I would need to run the
create-release.sh script SIX TIMES!

This contributes to issue #3