Refactor build script to build standard and addon pack types at once
Pulled the pack generation code into a function so that I can call that
code multiple times within the script without repeating myself. Now the
script generates both the standard edition and the addon pack during
the same run.

I also made some minor formatting changes to the text that is shown to
the user and changed some variable names to better separate the globals
from the locals within the function, to try to prevent future errors.

I updated the readme file to include a blurb about addon packs.  Also
added "and feature requests" to the section about bug reports, which
isn't related to this issue, but I happened to see it while editing, and
it is what it is.

This finally implements issue #3