Implement Mojang's new supported_formats field
In 1.20.2, Mojang added the field "supported_formats" to the list of
possible pack.mcmeta fields. It should allow pack creators like myself
to declare to the game that our packs are compatible with a wide range
of versions.

To this end, I reverted the "pack_format" specification to 15, and added
"supported_formats" on top of that.

I went back to 15 instead of leaving it at 18 because format 15 was the
last version of the game that didn't recognize the new range (barring
a few snapshots that hardly anyone will ever play.) This means that when
you include this pack on a 1.20 or 1.20.1 world, it will read
"pack_format:15" and declare that the pack is compatible.

I didn't go back prior to 15 because then Minecraft version 1.20 would
tell you that the pack was made for an older version of the game, while
1.20.2 and above would tell you it was compatible.  This was not ideal.

So now, the pack's initial compatibility is 1.20, and due to the range
that I've put in, the pack shouldn't complain about incompatibility
until Mojang goes above double-digits with the format number.

This may be a bad decision, because let's say next year they decide to
completely revamp the way sound packs work, my packs would still be
considered "compatible" even if they're not. This is a bit
disconcerting, but I'm lazy and I have no interest in pushing a release
simply to update a single number in a single text file just so the pack
can be not red in the latest snapshot or whatever.

This implements issue #29