Add new sounds to adult edition from audio captured on 2023.06.10
Five new sounds added in this commit, including 4 villager "trade"
sounds and a villager "yes" sound.

I used my sound-pack-indexer to auto-generate the JSON records and copy
in the files, so it added whitespace to every single line in the
sounds.json file.  This basically marks my first real-world test of the
indexer concept.  So far, it's working ok.

Contributes to issue #10