Add new sounds to uwu edition from audio captured on 2023.06.10
1 witch "ambient" sound added in this pack.

I used my sound-pack-indexer to auto-generate the JSON records and copy
in the files, so it added whitespace to every single line in the
sounds.json file. This basically marks my first real-world test of the
indexer concept. So far, it's working ok. One glitch I've encountered
might be that it doesn't sort the entries alphabetically.  I'll look
into that later.

Implements issue #10