Remove witch sounds from Adult edition that are duplicates of Loaded
This was done because it is now possible to layer multiple packs
together.  If I hadn't removed them, there would be two records in
memory for each of the duplicated sounds, making them effectively
weight = 2 (or double the chance of playing over any unique witch
sounds in the pack.)

Furthermore, there are no "adult" overtones in any of the duplicated
sounds anyway (except the one that is left), so it's probably a good
thing to remove them.

As far as I can tell, the Uwu edition already has unique witch sounds.

Implements issue #34