Make sure that replace:true in build script works in various situations
After a change to Sound Pack Indexer, spaces were added after the colon
in "replace":true, which led to the build script not being able to find
the text to replace. Another change that happened is that occasionally
the Indexer app will put "replace": true on the bottom of the list
rather than at the top where one would expect "r" to come before "s",
so as a result, "replace": true would sometimes not be followed by a
comma.

I added regular expression syntax to remove zero or more whitespace
from between the colon and the word "true"