Add code to clear the temp dir when finished building an edition
I was copying in the files every time the build_pack function was
called, but never clearing the temp directory after the edition was
completely packed. This led to each subsequent edition having all the
files of the previous. Uwu edition then became the largest zip file, at
30MB! I never noticed this in my playtesting because Minecraft can only
play files it has json records for.

Now the files are copied once per edition (the addon pack shares the
same files, so no need to copy those in again), and once both packs
from each edition are produced, it clears out the temp folder for the
next edition.

Uwu went from 30MB to 2.7! The process is also a bit faster.